Post output to Google doc
This will allow to update the dashboard automatically, using heroku scheduler.
This prepares a batch request to send 100 rows at a time to circumvent Google
API's rate limiting. This also parses alerts better so no parsing needs to be
done in the excel sheet.

The authentication is done through the Google service account govuk-2ndline,
which allows server to server operations.